Create 2Sum.cpp
The "Two Sum" problem is a well-known and fundamental problem in computer science and data structures. It can be described as follows:

Problem Statement:
Given an array of integers and a target sum, the Two Sum problem asks whether there are two distinct elements in the array that add up to the target sum. If such elements exist, the algorithm should return their indices.